+++ src/main/java/com/chrisrm/idea/MTConfig.java
@@ -41,7 +41,7 @@ import com.intellij.openapi.components.Storage;
 import com.intellij.ui.ColorUtil;
 import com.intellij.util.ObjectUtils;
 import com.intellij.util.xmlb.XmlSerializerUtil;
-import io.acari.DDLC.actions.ChibiLevel;
+import io.acari.DDLC.chibi.ChibiLevel;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.json.JSONException;
