+++ src/main/java/io/acari/DDLC/actions/AllChibiAction.kt
@@ -1,3 +1,5 @@
 package io.acari.DDLC.actions
 
+import io.acari.DDLC.chibi.ChibiLevel
+
 class AllChibiAction: BaseChibiAction(ChibiLevel.OVER9000)
