+++ src/main/java/io/acari/DDLC/actions/BaseChibiAction.kt
@@ -2,6 +2,7 @@ package io.acari.DDLC.actions
 
 import com.chrisrm.idea.actions.themes.BaseThemeAction
 import com.intellij.openapi.actionSystem.AnActionEvent
+import io.acari.DDLC.chibi.ChibiLevel
 import io.acari.DDLC.chibi.ChibiOrchestrator
 
 open class BaseChibiAction(private val chibiLevel: ChibiLevel) : BaseThemeAction() {
