+++ src/main/java/io/acari/DDLC/actions/SomeChibiAction.kt
@@ -1,3 +1,5 @@
 package io.acari.DDLC.actions
 
+import io.acari.DDLC.chibi.ChibiLevel
+
 class SomeChibiAction: BaseChibiAction(ChibiLevel.ON)
