+++ src/main/java/io/acari/DDLC/chibi/ChibiLevel.kt
@@ -1,4 +1,4 @@
-package io.acari.DDLC.actions
+package io.acari.DDLC.chibi
 
 enum class ChibiLevel {
     OFF, ON, OVER9000
